pyproject.toml
[project]
name = "beyond-ralph"
version = "0.1.0"
description = "True Autonomous Coding - Multi-agent orchestration for Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "jaymzjulian", email = "jaymzjulian@users.noreply.github.com"}
]
keywords = ["claude", "ai", "autonomous", "coding", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9",
    "aiofiles>=23.0",
    "tenacity>=8.0",
    "psutil>=5.9",

    # API/Backend Testing
    "httpx>=0.25",
    "responses>=0.24",

    # Web UI Testing
    "playwright>=1.40",

    # CLI Testing
    "pexpect>=4.9",

    # Desktop/Graphics Testing
    "pillow>=10.0",
    "opencv-python>=4.8",

    # Research capabilities
    "beautifulsoup4>=4.12",
    "lxml>=5.0",

    # Code Review Agent - Linting & Security
    "semgrep>=1.50",         # Multi-language security rules
    "bandit>=1.7",           # Python security linter
    "safety>=2.3",           # Dependency vulnerability checker
    "vulture>=2.10",         # Dead code detection
    "radon>=6.0",            # Code complexity metrics
    "detect-secrets>=1.4",   # Secrets detection
]

[project.optional-dependencies]
desktop = [
    # GUI automation (optional, may require system deps)
    "pyautogui>=0.9",
]

mobile = [
    # Mobile testing (discovered/installed as needed)
    "Appium-Python-Client>=3.0",
]

dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "mypy>=1.5",
    "ruff>=0.1",
    "pre-commit>=3.4",
]

[project.scripts]
beyond-ralph = "beyond_ralph.cli:main"
br-quota = "beyond_ralph.utils.quota_checker:main"
br-live-tests = "beyond_ralph.testing.claude_driver:main"

[project.entry-points."claude_code.skills"]
beyond-ralph = "beyond_ralph.skills:register_skills"

[project.entry-points."claude_code.hooks"]
beyond-ralph = "beyond_ralph.hooks:register_hooks"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beyond_ralph"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["beyond_ralph"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--cov=beyond_ralph",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=90",  # Target is 90% coverage
]
markers = [
    "integration: marks tests as integration tests (may call live services)",
    "live: marks tests that spawn real Claude CLI sessions (require quota)",
    "slow: marks tests that take longer than 30 seconds to run",
]

[tool.coverage.run]
source = ["src/beyond_ralph"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
